"-"제거하기
