+++ api/Kiwoom.py
@@ -71,9 +71,11 @@ class Kiwoom(QAxWidget):
         return code_name
 
     def get_price_data(self, code):
+        # 분봉 데이터 조회
         self.dynamicCall("SetInputValue(QString, QString)", "종목코드", code)
+        self.dynamicCall("SetInputValue(QString, QString)", "틱범위", 3)
         self.dynamicCall("SetInputValue(QString, QString)", "수정주가구분", "1")
-        self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10081_req", "opt10081", 0, "0001")
+        self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10080_req", "opt10080", 0, "0101")
 
         self.tr_event_loop.exec_()
 
@@ -81,12 +83,32 @@ class Kiwoom(QAxWidget):
 
         while self.has_next_tr_data:
             self.dynamicCall("SetInputValue(QString, QString)", "종목코드", code)
+            self.dynamicCall("SetInputValue(QString, QString)", "틱범위", 3)
             self.dynamicCall("SetInputValue(QString, QString)", "수정주가구분", "1")
-            self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10081_req", "opt10081", 2, "0001")
+            self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10080_req", "opt10080", 2, "0101")
             self.tr_event_loop.exec_()
 
             for key, val in self.tr_data.items():
-                ohlcv[key] += val
+                 ohlcv[key] += val
+
+
+        # 일봉데이터 조회
+        # self.dynamicCall("SetInputValue(QString, QString)", "종목코드", code)
+        # self.dynamicCall("SetInputValue(QString, QString)", "수정주가구분", "1")
+        # self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10081_req", "opt10081", 0, "0001")
+
+        # self.tr_event_loop.exec_()
+
+        # ohlcv = self.tr_data
+
+        # while self.has_next_tr_data:
+        #     self.dynamicCall("SetInputValue(QString, QString)", "종목코드", code)
+        #     self.dynamicCall("SetInputValue(QString, QString)", "수정주가구분", "1")
+        #     self.dynamicCall("CommRqData(QString, QString, int, QString)", "opt10081_req", "opt10081", 2, "0001")
+        #     self.tr_event_loop.exec_()
+
+            # for key, val in self.tr_data.items():
+            #     ohlcv[key] += val
 
         df = pd.DataFrame(ohlcv, columns=['open', 'high', 'low', 'close', 'volume'], index=ohlcv['date'])
 
@@ -102,7 +124,7 @@ class Kiwoom(QAxWidget):
         else:
             self.has_next_tr_data = False
 
-        if rqname == "opt10081_req":
+        if rqname == "opt10080_req":
             ohlcv = {'date': [], 'open': [], 'high': [], 'low': [], 'close': [], 'volume': []}
 
             for i in range(tr_data_cnt):
@@ -113,11 +135,12 @@ class Kiwoom(QAxWidget):
                 close = self.dynamicCall("GetCommData(QString, QString, int, QString", trcode, rqname, i, "현재가")
                 volume = self.dynamicCall("GetCommData(QString, QString, int, QString", trcode, rqname, i, "거래량")
 
+
                 ohlcv['date'].append(date.strip())
-                ohlcv['open'].append(int(open))
-                ohlcv['high'].append(int(high))
-                ohlcv['low'].append(int(low))
-                ohlcv['close'].append(int(close))
+                ohlcv['open'].append(int(open.strip().replace("-", "")))
+                ohlcv['high'].append(int(high.strip().replace("-", "")))
+                ohlcv['low'].append(int(low.strip().replace("-", "")))
+                ohlcv['close'].append(int(close.strip().replace("-", "")))
                 ohlcv['volume'].append(int(volume))
 
             self.tr_data = ohlcv
